+++ app/src/main/java/org/simple/clinic/home/patients/PatientsTabScreen.kt
@@ -209,7 +209,7 @@ class PatientsTabScreen : BaseScreen<
 
   override fun createInit() = PatientsInit(
       isNotifyAppUpdateAvailableV2Enabled = features.isEnabled(NotifyAppUpdateAvailableV2),
-      isMonthlyDrugStockReportReminderEnabled = features.isEnabled(MonthlyDrugStockReportReminder)
+      isMonthlyDrugStockReportReminderEnabledInIndia = features.isEnabled(MonthlyDrugStockReportReminder) && country.isoCountryCode == Country.INDIA
   )
 
   override fun createEffectHandler(viewEffectsConsumer: Consumer<PatientsTabViewEffect>) = effectHandlerFactory.create(
@@ -230,7 +230,8 @@ class PatientsTabScreen : BaseScreen<
     homeIllustration.setImageResource(illustrationResourceId())
     simpleVideoIllustration.setImageResource(videoIllustrationResourceId())
 
-    if (features.isEnabled(MonthlyDrugStockReportReminder)) {
+    val isMonthlyDrugStockReminderEnabledInIndia = features.isEnabled(MonthlyDrugStockReportReminder) && country.isoCountryCode == Country.INDIA
+    if (isMonthlyDrugStockReminderEnabledInIndia) {
       drugStockNotificationScheduler.schedule()
     }
   }